+++ lib/qc/modules/qc_tester/audit/providers/audit_submission_provider.dart
@@ -4,39 +4,60 @@ import 'package:core/core.dart';
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/audit/resources/new_audit_response.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/calculator_service.dart';
 import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/device_status_response.dart';
-import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/qc_calculator_service.dart';
+import 'package:flutter_trc/qc/modules/qc_tester/calculator/resources/media_submit_request.dart';
 import 'package:provider/provider.dart';
 
 import '../resources/audit_service.dart';
 
-class AuditQuestionSubmitProvider extends CshChangeNotifier {
+class AuditQuestionSubmitProvider extends CalculatorServiceInitProvider {
   List<int>? manualQuestionListIds;
 
   static AuditQuestionSubmitProvider of(BuildContext context, {bool listen = true}) {
     return Provider.of<AuditQuestionSubmitProvider>(context, listen: listen);
   }
 
-  AuditQuestionSubmitProvider();
-
   //Submit Audit question responses
-  Future<bool> submitAuditQuestion(String scannedBarcode, Map<String, dynamic> postData) {
+  Future<bool> submitAuditQuestion(
+    String scannedBarcode,
+    Map<String, dynamic> postData,
+    List<MediaSubmitRequest>? mediaList,
+  ) {
     var completer = Completer<bool>();
     try {
       AuditDataServices.submitAutQuestionResponses(
         scannedBarcode,
         postData,
         manualAuditQuestionIds: manualQuestionListIds,
+        service: service.getService(),
       ).listen((event) {
         if (event != null) {
           completer.complete(true);
+          _submitDeviceImages(mediaList, scannedBarcode);
         }
       }, onError: (error) {
-        String errorMessage = ApiErrorHelper.getErrorMessage(error) ?? "Some error occurred";
-
-        completer.completeError(errorMessage);
+        completer.completeError(error);
       }, onDone: () {
         notifyListeners();
+      });
+    } catch (e) {
+      completer.completeError(e.toString());
+    }
+    return completer.future;
+  }
+
+  void _submitDeviceImages(List<MediaSubmitRequest>? mediaList, String? deviceBarcode) {
+    service.submitDeviceMedia(mediaList, deviceBarcode).listen((event) {}, onError: (error) {});
+  }
+
+  Future<bool> checkDeviceAuditResult(String scannedBarcode, Map<String, dynamic> postData) {
+    var completer = Completer<bool>();
+    try {
+      AuditDataServices.checkIsTestingPass(scannedBarcode, postData, service: service.getService()).listen((event) {
+        completer.complete(event?.isTestingPass);
+      }, onError: (error) {
+        completer.completeError(error);
       });
     } catch (e) {
       completer.completeError(e.toString());
@@ -54,7 +75,7 @@ class AuditQuestionSubmitProvider extends CshChangeNotifier {
           continue;
         }
 
-        if (Validator.isNullOrEmpty(deviceStatusResponse.trcStatus) &&
+        if (Validator.isNullOrEmpty(deviceStatusResponse.status) &&
             Validator.isListNullOrEmpty(deviceStatusResponse.salesChannels)) {
           await Future.delayed(const Duration(seconds: 4));
           continue;
@@ -63,8 +84,8 @@ class AuditQuestionSubmitProvider extends CshChangeNotifier {
         String deviceStatus;
         String subTitle = deviceStatusResponse.stockAge != null ? "Stock Age - ${deviceStatusResponse.stockAge}, " : "";
 
-        if (!Validator.isNullOrEmpty(deviceStatusResponse.trcStatus)) {
-          deviceStatus = subTitle + deviceStatusResponse.trcStatus!;
+        if (!Validator.isNullOrEmpty(deviceStatusResponse.status)) {
+          deviceStatus = subTitle + deviceStatusResponse.status!;
         } else {
           deviceStatus = subTitle + deviceStatusResponse.salesChannels!.join(",");
         }
@@ -80,7 +101,7 @@ class AuditQuestionSubmitProvider extends CshChangeNotifier {
 
   Future<DeviceStatusResponse?> _callingDeviceStatusApi(String deviceBarcode) {
     var completer = Completer<DeviceStatusResponse?>();
-    QcCalculatorService().getDeviceStatus(deviceBarcode).listen((event) {
+    service.getDeviceStatus(deviceBarcode).listen((event) {
       completer.complete(event);
     }, onError: (error) {
       completer.completeError(error);