+++ lib/qc/modules/qc_tester/calculator_media_capture/components/calculator_media_capture_component.dart
@@ -29,11 +29,11 @@ class CalculatorMediaCaptureComponent extends StatelessComponent<NoneConfigModel
         return ChangeNotifierProvider<CalculatorMediaCaptureProvider>(
           create: (_) => CalculatorMediaCaptureProvider(
             model.deviceBarcode ?? "",
-            model.isComingFromCalJourney,
+            model.journeyType,
             categoryId: model.categoryId,
           ),
           lazy: false,
-          child: const CalculatorMediaCaptureWidget(),
+          child: CalculatorMediaCaptureWidget(onMediaListUpdated: model.onMediaListUpdated),
         );
       },
     );