+++ lib/qc/modules/qc_tester/calculator_media_capture/resources/journey_type.dart
@@ -1,0 +1,10 @@
+enum JourneyType {
+  generic,
+  testing,
+  audit;
+
+  bool get isTesting => this == JourneyType.testing;
+  bool get isAudit => this == JourneyType.audit;
+  bool get isGeneric => this == JourneyType.generic;
+
+}