+++ lib/src/common/provider/qc_trc_service_init_provider.dart
@@ -11,7 +11,7 @@ class QcTrcServiceInitProvider extends CshChangeNotifier {
     initService();
   }
 
-  Future<void> initService() async {
+  void initService() {
     var isLoginFromQc = isLoginFromQC();
     if (Validator.isTrue(isLoginFromQc)) {
       service = QcService();